add recv_post after sending the message
